TravelingSalesmanExact.jl

TravelingSalesmanExact.ATT

TravelingSalesmanExact.euclidean_distance

TravelingSalesmanExact.find_cycle

TravelingSalesmanExact.get_ATT48_cities

TravelingSalesmanExact.get_cycles

TravelingSalesmanExact.get_default_optimizer

TravelingSalesmanExact.get_optimal_tour

TravelingSalesmanExact.get_optimal_tour

TravelingSalesmanExact.plot_cities

TravelingSalesmanExact.plot_tour

TravelingSalesmanExact.remove_cycles!

TravelingSalesmanExact.set_default_optimizer!

TravelingSalesmanExact.simple_parse_tsp

TravelingSalesmanExact.get_optimal_tour — Function

get_optimal_tour(
    cost::AbstractMatrix,
    optimizer = get_default_optimizer();
    verbose::Bool = false,
    symmetric::Bool = issymmetric(cost),
    lazy::Bool = true,
)

Solves the travelling salesman problem for a square cost matrix using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path.

The second argument is mandatory if a default optimizer has not been set (via set_default_optimizer). This argument should be a function which creates an optimizer, e.g.

    get_optimal_tour(cities, GLPK.Optimizer)

There are three boolean optional keyword arguments:

verbose indicates whether or not to print lots of information as the algorithm proceeds.

symmetric indicates whether or not the cost matrix is symmetric (the default is to check via issymmetric)

lazy indicates whether lazy constraints should be used (which requires a compatible solver).

TravelingSalesmanExact.get_optimal_tour — Function

get_optimal_tour(
    cities::AbstractVector,
    optimizer = get_default_optimizer();
    verbose = false,
    distance = euclidean_distance,
    symmetric = true,
    lazy_constraints = false,
)

Solves the travelling salesman problem for a list of cities using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path. Optionally specify a distance metric.

The second argument is mandatory if a default optimizer has not been set (via set_default_optimizer). This argument should be a function which creates an optimizer, e.g.

get_optimal_tour(cities, GLPK.Optimizer)

There are three boolean optional keyword arguments:

verbose indicates whether or not to print lots of information as the algorithm proceeds.

symmetric indicates whether or not the distance metric used is symmetric (the default is to assume that it is)

lazy_constraints indicates whether lazy constraints should be used (which requires a compatible solver).

TravelingSalesmanExact.plot_cities — Method

plot_cities(cities)

Uses UnicodePlots's lineplot to make a plot of the tour of the cities in cities, in order (including going from the last city back to the first).

TravelingSalesmanExact.set_default_optimizer! — Method

set_default_optimizer(O)

Sets the default optimizer. For example,

using GLPK
set_default_optimizer(GLPK.Optimizer)

TravelingSalesmanExact.simple_parse_tsp — Method

simple_parse_tsp(filename; verbose = true)

Try to parse the ".tsp" file given by filename. Very simple implementation just to be able to test the optimization; may break on other files. Returns a list of cities for use in get_optimal_tour.

TravelingSalesmanExact.ATT — Method

ATT(city1, city2)

The ATT distance measure as specified in TSPLIB: https://www.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/tsp95.pdf.

TravelingSalesmanExact.euclidean_distance — Method

euclidean_distance(city1, city2)

The usual Euclidean distance measure.

TravelingSalesmanExact.find_cycle — Function

find_cycle(perm_matrix, starting_ind)

Returns the cycle in the permutation described by perm_matrix which includes starting_ind.

TravelingSalesmanExact.get_ATT48_cities — Method

get_ATT48_cities() -> Vector{Vector{Int}}

A simple helper function to get the problem data for the ATT48 TSPLIB problem.

Example

using TravelingSalesmanExact, GLPK
cities = TravelingSalesmanExact.get_ATT48_cities()
get_optimal_tour(cities, GLPK.Optimizer, distance = TravelingSalesmanExact.ATT)

TravelingSalesmanExact.get_cycles — Method

get_cycles(perm_matrix)

Returns a list of cycles from the permutation described by perm_matrix.

TravelingSalesmanExact.get_default_optimizer — Method

get_default_optimizer()

Gets the default optimizer, which is set by set_default_optimizer.

TravelingSalesmanExact.plot_tour — Method

show_tour(cities, perm_matrix)

Show a plot of the tour described by perm_matrix of the cities in the vector cities.

TravelingSalesmanExact.remove_cycles! — Method

remove_cycles!(model, tour_matrix)

Find the (non-maximal-length) cycles in the current solution tour_matrix and add constraints to the JuMP model to disallow them. Returns the number of cycles found.