Home

TravelingSalesmanExact.jl

TravelingSalesmanExact.ATT

TravelingSalesmanExact.euclidean_distance

TravelingSalesmanExact.find_cycle

TravelingSalesmanExact.get_cycles

TravelingSalesmanExact.get_optimal_tour

TravelingSalesmanExact.get_optimal_tour

TravelingSalesmanExact.plot_cities

TravelingSalesmanExact.plot_tour

TravelingSalesmanExact.remove_cycles!

TravelingSalesmanExact.simple_parse_tsp

TravelingSalesmanExact.get_optimal_tour — Method.

get_optimal_tour(cities::AbstractVector, with_optimizer; verbose = false, distance = euclidean_distance, symmetric = true)

Solves the travelling salesman problem for a list of cities using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path. Optionally specify a distance metric.

The second argument should be the result of a call to JuMP.with_optimizer, e.g.

get_optimal_tour(cities, with_optimizer(GLPK.Optimizer))

source

TravelingSalesmanExact.get_optimal_tour — Method.

get_optimal_tour(cost::AbstractMatrix, with_optimizer; verbose = false, symmetric = issymmetric(cost))

Solves the travelling salesman problem for a square cost matrix using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path.

The second argument should be the result of a call to JuMP.with_optimizer, e.g.

get_optimal_tour(cost, with_optimizer(GLPK.Optimizer))

source

TravelingSalesmanExact.plot_cities — Method.

plot_cities(cities)

Uses UnicodePlots's lineplot to make a plot of the tour of the cities in cities, in order (including going from the last city back to the first).

source

TravelingSalesmanExact.simple_parse_tsp — Method.

simple_parse_tsp(filename; verbose = true)

Try to parse the ".tsp" file given by filename. Very simple implementation just to be able to test the optimization; may break on other files. Returns a list of cities for use in get_optimal_tour.

source

TravelingSalesmanExact.ATT — Method.

ATT(city1, city2)

The ATT distance measure as specified in TSPLIB: https://www.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/tsp95.pdf.

source

TravelingSalesmanExact.euclidean_distance — Method.

euclidean_distance(city1, city2)

The usual Euclidean distance measure.

source

TravelingSalesmanExact.find_cycle — Function.

find_cycle(perm_matrix, starting_ind)

Returns the cycle in the permutation described by perm_matrix which includes starting_ind.

source

TravelingSalesmanExact.get_cycles — Method.

get_cycles(perm_matrix)

Returns a list of cycles from the permutation described by perm_matrix.

source

TravelingSalesmanExact.plot_tour — Method.

show_tour(cities, perm_matrix)

Show a plot of the tour described by perm_matrix of the cities in the vector cities.

source

TravelingSalesmanExact.remove_cycles! — Method.

remove_cycles!(model, tour_matrix)

Find the (non-maximal-length) cycles in the current solution tour_matrix and add constraints to the JuMP model to disallow them. Returns the number of cycles found.

source