Home

TSP_MIP.jl

TSP_MIP.ATT

TSP_MIP.euclidean_distance

TSP_MIP.find_cycle

TSP_MIP.get_cycles

TSP_MIP.get_optimal_tour

TSP_MIP.get_optimal_tour

TSP_MIP.plot_cities

TSP_MIP.plot_tour

TSP_MIP.remove_cycles!

TSP_MIP.simple_parse_tsp

TSP_MIP.get_optimal_tour — Method.

get_optimal_tour(cities::AbstractVector; verbose = true, distance = euclidean_distance, optimizer = GLPK.Optimizer)

Solves the travelling salesman problem for a list of cities using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path. Optionally specify a distance metric and an optimizer for JuMP.

source

TSP_MIP.get_optimal_tour — Method.

get_optimal_tour(cost::AbstractMatrix; verbose = true, optimizer = GLPK.Optimizer)

Solves the travelling salesman problem for a square cost matrix using JuMP by formulating a MILP using the Dantzig-Fulkerson-Johnson formulation and adaptively adding constraints to disallow non-maximal cycles. Returns an optimal tour and the cost of the optimal path. Optionally specify an optimizer for JuMP.

source

TSP_MIP.plot_cities — Method.

plot_cities(cities)

Uses UnicodePlots's lineplot to make a plot of the tour of the cities in cities, in order (including going from the last city back to the first).

source

TSP_MIP.simple_parse_tsp — Method.

simple_parse_tsp(filename; verbose = true)

Try to parse the ".tsp" file given by filename. Very simple implementation just to be able to test the optimization; may break on other files. Returns a list of cities for use in get_optimal_tour.

source

TSP_MIP.ATT — Method.

ATT(city1, city2)

The ATT distance measure as specified in TSPLIB: https://www.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/tsp95.pdf.

source

TSP_MIP.euclidean_distance — Method.

euclidean_distance(city1, city2)

The usual Euclidean distance measure.

source

TSP_MIP.find_cycle — Function.

find_cycle(perm_matrix, starting_ind)

Returns the cycle in the permutation described by perm_matrix which includes starting_ind.

source

TSP_MIP.get_cycles — Method.

get_cycles(perm_matrix)

Returns a list of cycles from the permutation described by perm_matrix.

source

TSP_MIP.plot_tour — Method.

show_tour(cities, perm_matrix)

Show a plot of the tour described by perm_matrix of the cities in the vector cities.

source

TSP_MIP.remove_cycles! — Method.

remove_cycles!(model, tour_matrix)

Find the (non-maximal-length) cycles in the current solution tour_matrix and add constraints to the JuMP model to disallow them. Returns the number of cycles found.

source